% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_get_all_dups.R
\name{epi_get_all_dups}
\alias{epi_get_all_dups}
\title{Get all duplicated rows including the originals

epi_get_all_dups() Get all elements which are repeated,
including the originals.}
\usage{
epi_get_all_dups(df = NULL, var = "", freq = 1)
}
\arguments{
\item{df}{A data.frame to extract duplicates from}

\item{var}{Variable/column name as a string to use to detect duplicate values}

\item{freq}{Frequency count, an integer. Defaults to 1 (so will return original
plus any repeating values)}
}
\value{
a data.frame with original and duplicated rows
}
\description{
base R duplicated() does not return the originals (duplicated - 1)
See for example https://stackoverflow.com/questions/16905425/find-duplicate-values-in-r
}
\note{
\preformatted{Creates a table with frequencies, checks those which have more than 1
}

(or value passed to freq and considers these as variables with duplicates)
and extracts them.
}
\examples{

\dontrun{
dim(df)
epi_head_and_tail(df, rows = 2, cols = 2)
# Get all duplicates:
check_dups <- epi_get_all_dups(df, 'var_id', 1)
dim(check_dups)
check_dups
}

}
\seealso{
\code{\link{duplicated}}
}
\author{
Antonio Berlanga-Taylor <\url{https://github.com/AntonioJBT/episcout}>
}
