% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_output_name.R
\name{epi_output_name}
\alias{epi_output_name}
\title{Create a name for plots and files to save}
\usage{
epi_output_name(input_name = "", suffix = ".tsv")
}
\arguments{
\item{input_name}{string to split, usually an input file name previously saved as an object}

\item{suffix}{any string you wish to add, default is '.tsv'}
}
\value{
string value
}
\description{
Generate a simple output name, given a string (filename) splits at the last '.', drops the current suffix and adds the one provided. If \code{input_name} lacks a dot, the whole string is used as is.
}
\examples{
\dontrun{
input_name <- "baseline_clean.RData"
svg(epi_output_name(input_name, "_histogram_obesity_1.svg"))
plot(x, y)
dev.off()
}

}
\author{
Antonio Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
