% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot_all.R
\name{epi_plot_bar_grid_save}
\alias{epi_plot_bar_grid_save}
\title{Save bar plot grids in parallel}
\usage{
epi_plot_bar_grid_save(
  plot_list,
  results_subdir,
  file_n = "plots_bar",
  per_file = 4,
  suffix = "pdf",
  n_cores = parallel::detectCores() - 1
)
}
\arguments{
\item{plot_list}{List of ggplot objects.}

\item{results_subdir}{Directory where files will be written.}

\item{file_n}{File name prefix. Defaults to "plots_bar".}

\item{per_file}{Number of plots per saved grid. Defaults to 4.}

\item{suffix}{File extension, e.g. "pdf".}

\item{n_cores}{Number of cores for parallel processing. Defaults to
\code{parallel::detectCores() - 1}.}
}
\value{
A vector of output file paths.
}
\description{
Split a list of plots into groups and save each group as a grid to disk in
parallel.
}
\examples{
\dontrun{
df <- data.frame(sex = factor(c("F", "M"), levels = c("F", "M")))
plots <- epi_plot_bar_list_parallel("sex", df, n_cores = 1)
dir <- tempdir()
epi_plot_bar_grid_save(
  plot_list = plots,
  results_subdir = dir,
  file_n = "my_plots",
  per_file = 1,
  suffix = "png",
  n_cores = 1
)
list.files(dir, pattern = "my_plots")
}
}
