% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot_all.R
\name{epi_plot_bar_list_parallel}
\alias{epi_plot_bar_list_parallel}
\title{Generate bar plots in parallel}
\usage{
epi_plot_bar_list_parallel(
  vars_to_plot,
  data,
  custom_palette = NULL,
  n_cores = parallel::detectCores() - 1
)
}
\arguments{
\item{vars_to_plot}{Character vector of factor variables to plot.}

\item{data}{A data frame containing \code{vars_to_plot}.}

\item{custom_palette}{Optional vector of colours passed to
\code{\link[=epi_plot_bar]{epi_plot_bar()}}.}

\item{n_cores}{Number of cores for parallel processing. Defaults to
\code{parallel::detectCores() - 1}.}
}
\value{
A named list of ggplot objects.
}
\description{
Create a named list of bar plots for the specified factor variables using
\code{\link[=epi_plot_bar]{epi_plot_bar()}} in parallel.
}
\examples{
\dontrun{
df <- data.frame(
  sex = factor(sample(c("F", "M"), 20, replace = TRUE)),
  group = factor(sample(letters[1:3], 20, replace = TRUE))
)
plot_list <- epi_plot_bar_list_parallel(
  vars_to_plot = c("sex", "group"),
  data = df,
  custom_palette = c("#1b9e77", "#d95f02"),
  n_cores = 1
)
plot_list[[1]]
}
}
