% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot_en_masse.R
\name{epi_plot_en_masse}
\alias{epi_plot_en_masse}
\title{Batch processing for plots into figures for many columns}
\usage{
epi_plot_en_masse(df = NULL, file_prefix = "", plot_type = "svg",
  var_type = "numeric", plot_step = 6, ...)
}
\arguments{
\item{df}{Dataframe with variables to plot.}

\item{file_prefix}{Prefix to use for filenames. Default is ''.}

\item{plot_type}{Plot type to generate, this is passed to cowplot::save_plot.
Default is 'svg'.}

\item{var_type}{Type of variables to plot, numeric, integer or factor.
Default is 'numeric'.}

\item{plot_step}{Number of plots per page (file). Default is 6.}

\item{...}{Arguments to pass to epi_utils_multicore().}
}
\value{
Nothing, saves plots to disk.
}
\description{
epi_plot_en_masse() takes a data.frame and column type,
creates multi-plot figures for all and saves to disk.
Useful when plotting tens to hundreds of plots.
By default runs in parallel.
}
\note{
epi_utils_multicore() will run with defaults (max cores - 1 and multiprocess).
You may need to experiment first to see how many plots fit per page.
}
\examples{

\dontrun{

# Adjust labels and generate plots:
lab <- stringr::str_replace_all(string = i, pattern = '_', replacement = ' ')
lab <- stringr::str_to_upper(lab)
}

}
\seealso{
\code{\link{epi_plot_list}},
\code{\link{epi_plots_to_grid}},
\code{\link{epi_plot_cow_save}},
\code{\link{epi_utils_multicore}}.
}
\author{
Antonio Berlanga-Taylor <\url{https://github.com/AntonioJBT/episcout}>
}
