% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot_grid_size.R
\name{epi_plot_grid_size}
\alias{epi_plot_grid_size}
\title{Calculate number of plots to pass to grid for plotting}
\usage{
epi_plot_grid_size(plot_list = NULL, max_cols = 2, max_rows = 6)
}
\arguments{
\item{plot_list}{String of plot names.}

\item{max_cols}{Maximum number of columns in grid (page) Default is 2.}

\item{max_rows}{Maximum number of rows in grid. Default is 6.}
}
\value{
a list object containing nrow and ncols to use per grid
}
\description{
Get a size for the grid plot for multi-plot figures.
Preferably to have a grid size at most 2 cols by 3 rows for 6 plots total.
Can pass more with max_cols and max_rows though max_rows should be 6 when
max_cols is 1
}
\note{
It is probably easier to define these variables manually unless you are
plotting many variables.

See example in \code{\link{epi_plot_cow_save}} and ggplot2 wrappers epi_plot_*().
}
\seealso{
\code{\link{epi_plot_list}},
\code{\link{epi_plots_to_grid}},
\code{\link{epi_plot_cow_save}}.
}
\author{
Antonio Berlanga-Taylor <\url{https://github.com/AntonioJBT/episcout}>
}
\keyword{internal}
