% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot_hist.R
\name{epi_plot_hist}
\alias{epi_plot_hist}
\title{Histogram wrapper function using ggplot2}
\usage{
epi_plot_hist(df = NULL, var_x = "", ...)
}
\arguments{
\item{df}{data.frame with x var to plot}

\item{var_x}{Variable to plot on x-axis, pass as a string.}

\item{...}{passes arguments to ggplot2::geom_histogram()}
}
\value{
Prints a ggplot2 histogram
}
\description{
epi_plot_hist()
}
\note{
For other options, save as object and build on the layers. var_x is passed to ggplot2 aesthetics using tidy evaluation
}
\examples{
\dontrun{
library(ggplot2)
set.seed(12345)
n <- 20
df <- data.frame(
  var_id = rep(1:(n / 2), each = 2),
  var_to_rep = rep(c("Pre", "Post"), n / 2),
  x = rnorm(n),
  y = rbinom(n, 1, 0.50),
  z = rpois(n, 2)
)
df
df$x # continuous variable
my_hist_plot <- epi_plot_hist(df, "x") # pass with quotes as using ggplot2::aes_string()
my_hist_plot
# Change the bins:
my_hist_plot <- epi_plot_hist(df, "x", breaks = seq(-3, 3, by = 1))
my_hist_plot
# Add titles and axis names:
my_hist_plot <- my_hist_plot +
  labs(title = "Histogram for X") +
  labs(x = "X", y = "Count")
my_hist_plot
# Add axis limits:
my_hist_plot <- my_hist_plot +
  xlim(c(-4, 4)) +
  ylim(c(0, 10))
my_hist_plot
# Histogram with density curve:
my_hist_plot <- my_hist_plot + geom_density(col = 2)
my_hist_plot
# Histogram overlaid with kernel density curve:
# https://ggplot2.tidyverse.org/reference/geom_histogram.html
my_hist_plot <- my_hist_plot +
  # Density instead of count on y-axis:
  geom_histogram(aes(y = ..density..),
    binwidth = 0.5,
    colour = "black",
    fill = "white"
  ) +
  geom_density(alpha = 0.2, fill = "#FF6666") + # Overlay with transparent density plot
  ylab("Density")
my_hist_plot
}

}
\seealso{
\code{\link{epi_plot_list}}, \code{\link{epi_plots_to_grid}}, \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_histogram}}, \code{\link{epi_plot_cow_save}}.
}
\author{
Antonio Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
