% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot_list.R
\name{epi_plot_list}
\alias{epi_plot_list}
\title{Lists to hold plots}
\usage{
epi_plot_list(vars_to_plot = NULL)
}
\arguments{
\item{vars_to_plot}{string with variable names or possibly labels}
}
\value{
a list object
}
\description{
Create a list to hold plots for later plotting (in a multi- plot figure for example). Allows maximum 8 plots per list holder.
}
\note{
See example in \code{\link{epi_plot_cow_save}} and ggplot2 wrappers epi_plot_*().
}
\seealso{
\code{\link{epi_plots_to_grid}}, \code{\link{epi_plot_cow_save}}.
}
\author{
Antonio Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
