% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot_parallel.R
\name{epi_plot_parallel}
\alias{epi_plot_parallel}
\alias{epi_plot_save_parallel}
\title{Generate plots in parallel}
\usage{
epi_plot_parallel(
  df,
  vars_to_plot = NULL,
  var_type = "numeric",
  plot_fun = epi_plot_hist,
  num_cores = NULL,
  future_plan = "multisession",
  ...
)

epi_plot_save_parallel(
  plot_list,
  file_prefix = "",
  plot_type = "svg",
  plot_step = 6,
  num_cores = NULL,
  future_plan = "multisession",
  ...
)
}
\arguments{
\item{df}{A data frame containing variables to plot.}

\item{vars_to_plot}{Character vector of variable names to plot. If \code{NULL}, variables will be selected based on \code{var_type}.}

\item{var_type}{Type of variables to select when \code{vars_to_plot} is \code{NULL}. One of "numeric", "integer" or "factor".}

\item{plot_fun}{Function used to generate each plot. Defaults to \code{\link[=epi_plot_hist]{epi_plot_hist()}}.}

\item{num_cores}{Number of cores to use. Passed to \code{\link[=epi_utils_multicore]{epi_utils_multicore()}}. Defaults to all available minus one.}

\item{future_plan}{Strategy for parallel execution. See \code{\link[future:plan]{future::plan()}}.}

\item{...}{Additional arguments passed to \code{\link[=epi_utils_multicore]{epi_utils_multicore()}}.}

\item{plot_list}{Named list of plots to save.}

\item{file_prefix}{Prefix used to create file names.}

\item{plot_type}{File type passed to \code{\link[cowplot:save_plot]{cowplot::save_plot()}}.}

\item{plot_step}{Number of plots per file.}
}
\value{
A named list of plots with an attribute \code{workers} reporting the number of workers used.

A character vector of file names with an attribute \code{workers} reporting the number of workers used.
}
\description{
\code{epi_plot_parallel()} creates a list of plots for selected variables using a parallel backend. The default plotting function is \code{\link[=epi_plot_hist]{epi_plot_hist()}}, but any function returning a ggplot object and accepting \code{df} and a variable name can be supplied.

\code{epi_plot_save_parallel()} saves plots produced by \code{\link[=epi_plot_parallel]{epi_plot_parallel()}} to disk in parallel.
}
\examples{
\dontrun{
plots <- epi_plot_parallel(mtcars,
  vars_to_plot = c("mpg", "disp"),
  num_cores = 2
)
}

\dontrun{
plots <- epi_plot_parallel(mtcars,
  vars_to_plot = c("mpg", "disp"),
  num_cores = 2
)
files <- epi_plot_save_parallel(plots, file_prefix = "plots", plot_type = "png")
}

}
\seealso{
\code{\link[=epi_plot_save_parallel]{epi_plot_save_parallel()}}, \code{\link[=epi_utils_multicore]{epi_utils_multicore()}}

\code{\link[=epi_plot_parallel]{epi_plot_parallel()}}, \code{\link[=epi_plots_to_grid]{epi_plots_to_grid()}}, \code{\link[=epi_plot_cow_save]{epi_plot_cow_save()}}
}
