% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plots_to_grid.R
\name{epi_plots_to_grid}
\alias{epi_plots_to_grid}
\title{Send a list of plots to a grid object}
\usage{
epi_plots_to_grid(
  plot_list = NULL,
  align = "hv",
  axis = "lrtb",
  label_size = 12,
  ncol = NULL,
  nrow = NULL,
  ...
)
}
\arguments{
\item{plot_list}{List of plots to be arranged into the grid.}

\item{align}{cowplot vertical and/or horizontal alignment. Default is 'hv'.}

\item{axis}{Align left, right, top or bottom and order. Default is 'lrtb'.}

\item{label_size}{Default is 12.}

\item{ncol}{Number of columns in the plot grid.}

\item{nrow}{Number of rows in the plot grid.}

\item{...}{Pass any other parameter from plot_grid()}
}
\value{
a cowplot grid object
}
\description{
A light wrapper for cowplot::plot_grid(). Send a list of plots to a grid for multi-plot figures. Makes assumptions and hard-codes preferences. All options are passed to plot_grid().
}
\note{
See example in \code{\link{epi_plot_cow_save}} and ggplot2 wrappers epi_plot_*().
}
\seealso{
\code{\link{epi_plot_list}}, \code{\link{epi_plot_cow_save}}, \code{\link[cowplot]{plot_grid}}, \code{\link[cowplot]{save_plot}}.
}
\author{
Antonio Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
