% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_read.R
\name{epi_read}
\alias{epi_read}
\title{Read files with a consistent convenience function}
\usage{
epi_read(
  input_name = "",
  na.strings = c(-Inf, "NULL", NULL, ".", "", "NA", "NaN", NA, "<NA>"),
  header = TRUE,
  stringsAsFactors = FALSE,
  strip.white = TRUE,
  ...
)
}
\arguments{
\item{input_name}{file to read as a string}

\item{na.strings}{as for data.table::fread() with extended defaults}

\item{header}{as for data.table::fread(), default is TRUE}

\item{stringsAsFactors}{as for data.table::fread(), default is also FALSE}

\item{strip.white}{as for data.table::fread(), default is also TRUE}

\item{...}{pass any other data.table::fread() parameters}
}
\value{
A tibble
}
\description{
epi_read() reads files with a consistent convenience function. Wraps as.tibble and data.table's fread.
}
\note{
Other parameters as specified by data.table::fread() You are probably better off using the standard read.csv(), data.table::fread() or other functions unless you are reading in several similarly constructed files. Files are read with data.table first and then converted to a tibble. Columns will be read as integer, numeric or character only. White space is stripped. Strings are read as character only. Assumes the first row is a header.
}
\examples{
\dontrun{
super_data <- epi_read("super_data.tsv")
}

}
\seealso{
\code{\link[data.table]{fread}}, \code{\link[tibble]{as.tibble}}.
}
\author{
Antonio Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
