% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats_contingency_2x2.R
\name{epi_stats_2x2_aliases}
\alias{epi_stats_2x2_test}
\alias{epi_stats_2x2_cols}
\alias{epi_stats_2x2_all}
\title{Aliases for 2x2 contingency functions}
\usage{
epi_stats_2x2_test(...)

epi_stats_2x2_cols(...)

epi_stats_2x2_all(...)
}
\arguments{\item{...}{Arguments passed to the underlying \code{epi_stats_contingency_2x2_*} functions.}}
\description{These functions are wrappers around their \code{epi_stats_contingency_2x2_*} counterparts and are provided for backwards compatibility.}
