% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats_count_outliers.R
\name{epi_stats_count_outliers}
\alias{epi_stats_count_outliers}
\title{Count univariate outliers}
\usage{
epi_stats_count_outliers(num_vec = NULL, coef = 1.5, ...)
}
\arguments{
\item{num_vec}{Numeric vector to test.}

\item{coef}{Coefficient for outlier detection, default is 1.5}

\item{...}{Other parameters that can be passed to boxplot.stats().}
}
\value{
Returns the number of observations above the cut-off specified.
}
\description{
epi_stat_count_outliers() counts how many outliers a vector has using a univariate approach. Returns the number of observations that are greater than the cutoff. Outliers are detected with the Tukey method (above and below coef * IQR).
}
\details{
Returns 0 for empty or all-\code{NA} vectors, excludes non-finite values from Tukey calculations and raises an error when \code{coef} is negative. Setting \code{coef} to \code{0} disables outlier detection.
}
\note{
coef = 0 returns no outliers, see ?boxplot.stats An alternative, not implemented, is to consider those eg > 5 * SD
}
\examples{
\dontrun{
n <- 1000
df <- data.frame(
  var_id = rep(1:(n / 2), each = 2),
  var_to_rep = rep(c("Pre", "Post"), n / 2),
  x = rnorm(n),
  y = rbinom(n, 1, 0.50),
  z = rpois(n, 2)
)
epi_head_and_tail(df)
epi_stat_count_outliers(num_vec = df$x, coef = 0)
epi_stat_count_outliers(num_vec = df$x)
summary(df$x)
}

}
\seealso{
\code{\link[grDevices]{boxplot.stats}}
}
\author{
Antonio J Berlanga-Taylor <\url{https://github.com/AntonioJBT/episcout}>
}
