% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats_na_perc.R
\name{epi_stats_na_perc}
\alias{epi_stats_na_perc}
\title{Get counts and percentage of NAs}
\usage{
epi_stats_na_perc(df = NULL, margin = 2)
}
\arguments{
\item{df}{A dataframe with missing values}

\item{margin}{2 for columns, 1 for rows. Default is columns}
}
\value{
A dataframe object with counts and percentage of NAs for each column or row. Invalid margins raise a clear error.
}
\description{
epi_stats_na_perc() gets the total counts and percentage for missing values across rows or columns
}
\examples{
\dontrun{
library(mice)
dim(nhanes)
epi_head_and_tail(nhanes, cols = 4)
# Get summary of counts and percentages for missing values across columns:
na_cols <- epi_stats_na_perc(nhanes, margin = 2)
na_cols
# Get summary of counts and percentages for missing values across rows:
na_rows <- epi_stats_na_perc(nhanes, margin = 1)
epi_head_and_tail(na_rows, cols = 2)
summary(na_rows)
length(which(na_rows$na_counts == 0))
# which should be equal to:
length(which(complete.cases(nhanes)))
}

}
\seealso{
\code{\link[stats]{complete.cases}}
}
\author{
Antonio J Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
