% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot_km.R
\name{epi_survfit_to_df}
\alias{epi_survfit_to_df}
\title{Convert survfit object to data frame}
\usage{
epi_survfit_to_df(survfit_obj)
}
\arguments{
\item{survfit_obj}{A \code{survival::survfit} object.}
}
\value{
A data frame containing survival estimates and any strata variables.
}
\description{
\code{epi_survfit_to_df()} converts a \code{survival::survfit} object into a data frame suitable for plotting. Strata information is expanded into individual columns when present.
}
\examples{
\dontrun{
library(survival)
fit <- survfit(Surv(time, status) ~ sex, data = lung)
epi_survfit_to_df(fit)
}

}
