% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_utils_multicore.R
\name{epi_utils_multicore}
\alias{epi_utils_multicore}
\title{Setup a parallel run}
\usage{
epi_utils_multicore(
  num_cores = NULL,
  future_plan = "multisession",
  gc = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{num_cores}{Number of cores to use. Default cores - 1.}

\item{future_plan}{The strategy to use for plan(). Default is 'multisession'.}

\item{gc}{Garbage collection. Default is TRUE. Not all strategies use it though and cause an overhead.}

\item{verbose}{Prints out the settings for future::plan(). Default is TRUE.}

\item{...}{pass any further parameters accepted by future::plan().}
}
\value{
Nothing
}
\description{
epi_utils_multicore() setups a parallel function call using the package future.
}
\note{
num_cores is calculated as max(1, parallel::detectCores() - 1). If using sequential a warning will be emitted for workers. Not all strategies use gc(). See \code{\link[future]{plan}}.
}
\examples{
\dontrun{
epi_utils_multicore(num_cores = 1, future_plan = "sequential")
future_v \%<-\% {
  sum(1 + 2)
}
future_v
epi_utils_multicore(future_plan = "multisession")
future_v \%<-\% {
  sum(2 + 2)
}
future_v
}
}
\seealso{
\code{\link[future]{plan}}, \code{\link[parallel]{detectCores}}, \code{\link[doFuture]{registerDoFuture}}, \code{\link[future]{availableCores}}, \code{\link[future]{availableWorkers}}.
}
\author{
Antonio Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
