% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats_sum_vars.R
\name{epis_stats_chars}
\alias{epis_stats_chars}
\title{Summarize Character Variables}
\usage{
epis_stats_chars(df)
}
\arguments{
\item{df}{A dataframe containing character variables}
}
\value{
A tibble summarizing character variable statistics
}
\description{
Summarize Character Variables
}
