% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot_colour_palettes.R
\docType{data}
\name{palette_IMSS}
\alias{palette_IMSS}
\title{IMSS “official” colour palette}
\format{
An object of class \code{character} of length 10.
}
\usage{
palette_IMSS
}
\value{
A character vector of hex colour codes.
}
\description{
Vector of IMSS brand colours.
}
\keyword{datasets}
