% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_clean_curp.R
\name{print.epi_curp_audit}
\alias{print.epi_curp_audit}
\title{Print a CURP structural audit safely}
\usage{
\method{print}{epi_curp_audit}(x, ...)
}
\arguments{
\item{x}{An \code{epi_curp_audit} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a CURP structural audit safely
}
