% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_clean_curp.R
\name{str.epi_curp_audit}
\alias{str.epi_curp_audit}
\title{Display a safe structure summary for a CURP audit}
\usage{
\method{str}{epi_curp_audit}(object, ...)
}
\arguments{
\item{object}{An \code{epi_curp_audit} object.}

\item{...}{Unused.}
}
\value{
\code{object}, invisibly.
}
\description{
Display a safe structure summary for a CURP audit
}
