% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_suggests.R
\name{check_suggests}
\alias{check_suggests}
\title{Ensure suggested packages are installed}
\usage{
check_suggests(pkgs)
}
\arguments{
\item{pkgs}{Character vector of package names.}
}
\value{
Invisibly returns TRUE if all packages are installed. Throws an
error listing missing packages otherwise.
}
\description{
Checks that all packages listed are available for optional functionality.
}
\examples{
\dontrun{
check_suggests(c("ggplot2", "data.table"))
}
}
\keyword{internal}
