% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_clean.R
\name{epi_clean_add_colname_suffix}
\alias{epi_clean_add_colname_suffix}
\title{Compare two columns which may have duplicated information
TO DO: test and complete, not working
Test:
Change col names to baseline, time_1, time_2, etc.:}
\usage{
epi_clean_add_colname_suffix(df, id_col_num, suffix)
}
\description{
Compare two columns which may have duplicated information
TO DO: test and complete, not working
Test:
Change col names to baseline, time_1, time_2, etc.:
}
