% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_clean_cond_chr_fct.R
\name{epi_clean_cond_chr_fct}
\alias{epi_clean_cond_chr_fct}
\title{Check if column is character or factor}
\usage{
epi_clean_cond_chr_fct(col = NULL)
}
\arguments{
\item{col}{column header. This will be passed to
is.character(col) and is.factor(col)}
}
\value{
boolean TRUE/FALSE indicating whether column passed is type character
or factor
}
\description{
Check if column is character or factor (TRUE). Useful when
extracting columns from a large data frame using dplyr
(ie get all character and factor columns)
}
\examples{

\dontrun{
library(dplyr)
library(tibble)
col_chr <- data.frame('chr1' = rep(c('A', 'B')),
                      'chr2' = rep(c('C', 'D'))
             )
df_cont_chr <- as.tibble(cbind(df, col_chr))
df_cont_chr \%>\% select_if(~ epi_clean_cond_chr_fct(.))
epi_clean_cond_chr_fct(df_cont_chr[[2]]) # should be 'TRUE'
epi_clean_cond_chr_fct(df_cont_chr[, 'x']) # should be 'FALSE'
}

}
\seealso{
\code{\link[episcout]{epi_clean_cond_numeric}},
\code{\link[episcout]{epi_clean_cond_date}},
\code{\link[episcout]{epi_clean_class_to_factor}},
\code{\link[episcout]{epi_clean_count_classes}},
\code{\link[base]{is.character}},
\code{\link[base]{is.factor}}.
}
\author{
Antonio Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
