% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_clean_cond_date.R
\name{epi_clean_cond_date}
\alias{epi_clean_cond_date}
\title{Check if column is date using lubridate}
\usage{
epi_clean_cond_date(col = NULL)
}
\arguments{
\item{col}{column header. This will be passed to lubridate::is.Date(col) and lubridate::is.POSIXt(col)}
}
\value{
boolean TRUE/FALSE indicating whether column passed is type date
}
\description{
Check if column is date (TRUE). Useful when extracting columns from a large data frame using dplyr (ie get all date type columns)
}
\examples{
\dontrun{
set.seed(12345)
n <- 20
df <- data.frame(
  var_id = rep(1:(n / 2), each = 2),
  var_to_rep = rep(c("Pre", "Post"), n / 2),
  x = rnorm(n),
  y = rbinom(n, 1, 0.50),
  z = rpois(n, 2)
)
df_date <- df
df_date$date_col <- seq(as.Date("2018/1/1"), by = "year", length.out = 5)
df_date \%>\% select_if(~ epi_clean_cond_date(.))
epi_clean_cond_date(df_date[[2]]) # should be 'FALSE'
epi_clean_cond_date(df_date[, "date_col"]) # should be 'TRUE'
}

}
\seealso{
\code{\link{epi_clean_cond_numeric}}, \code{\link{epi_clean_class_to_factor}}, \code{\link{epi_clean_count_classes}}, \code{\link[base]{is.character}}, \code{\link[base]{is.factor}}.
}
\author{
Antonio Berlanga-Taylor <\url{https://github.com/AntonioJBT/episcout}>
}
