% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_clean_curp.R
\name{epi_clean_curp}
\alias{epi_clean_curp}
\title{Extraer datos del CURP en Mexico}
\usage{
epi_clean_curp(curp)
}
\arguments{
\item{curp}{Un vector de texto que contiene uno o mas CURPs. Cada CURP debe tener exactamente 18 caracteres.}
}
\value{
Un \code{tibble} con los componentes del CURP organizados en columnas:
\itemize{
\item \code{CURP}: El CURP original.
\item \code{PrimeraLetraApellidoPaterno}: Primera letra del apellido paterno.
\item \code{PrimeraVocalApellidoPaterno}: Primera vocal del apellido paterno.
\item \code{PrimeraLetraApellidoMaterno}: Primera letra del apellido materno.
\item \code{PrimeraLetraNombre}: Primera letra del nombre.
\item \code{AnoNacimiento}: Ano completo de nacimiento (e.g., "1990", "2002").
\item \code{MesNacimiento}: Mes de nacimiento (2 digitos).
\item \code{DiaNacimiento}: Dia de nacimiento (2 digitos).
\item \code{Sexo}: Genero registrado (H para hombre, M para mujer).
\item \code{EntidadFederativa}: Entidad federativa de nacimiento (codigo de 2 letras).
\item \code{PrimerasConsonantes}: Consonantes internas (3 caracteres).
\item \code{Homoclave}: Caracter alfanumerico asignado para evitar duplicados.
\item \code{DigitoVerificador}: Digito para validacion final.
}
}
\description{
Esta funcion descompone un CURP en sus componentes principales segun las posiciones definidas en su estructura.
}
\examples{
# Ejemplo de uso con un solo CURP
epi_clean_curp("GOMC900514HDFRLA07")

# Ejemplo con varios CURPs
curps <- c(
"GOMC900514HDFRLA07",  # Antes del 2000
"LOAM020715MMCRSR09"   # Despues del 2000
)

# Aplicar la funcion a un vector de CURPs
resultados <- purrr::map_dfr(curps, epi_clean_curp)

# Mostrar resultados
print(t(resultados))

}
