% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_clean_int_to_factor.R
\name{epi_clean_int_to_factor}
\alias{epi_clean_int_to_factor}
\title{Convert integer columns to factors}
\usage{
epi_clean_int_to_factor(dt)
}
\arguments{
\item{dt}{A \code{data.table} containing integer columns to convert.}
}
\value{
A \code{data.table} with integer columns converted to factors.
}
\description{
Converts columns of class integer to factors in a \code{data.table}, excluding
those of class \code{IDate}.
}
\examples{
library(data.table)
dt <- data.table(a = 1:3, b = as.IDate("2020-01-01") + 0:2)
epi_clean_int_to_factor(dt)
}
