% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_clean.R
\name{epi_clean_merge_nested_dfs}
\alias{epi_clean_merge_nested_dfs}
\title{Test:
df3 <- df
df3
df3 <- epi_clean_spread_repeated(df3, 'var_to_rep', 1)
df3
Recursively merge data frames that are stored as lists within a list
Flatten with eg purrr::flatten() if there is more than one level.
Assumes:
\itemize{
\item there are are 3 or more data frames to merge
\item there are no duplicates in any of the data frames
The function performs a full outer join with base R
merge(df1, df2, by = id_col, all = TRUE)
TO DO: if only two dfs in list, print message or do one merge
}}
\usage{
epi_clean_merge_nested_dfs(nested_list_dfs, id_col)
}
\description{
Test:
df3 <- df
df3
df3 <- epi_clean_spread_repeated(df3, 'var_to_rep', 1)
df3
Recursively merge data frames that are stored as lists within a list
Flatten with eg purrr::flatten() if there is more than one level.
Assumes:
\itemize{
\item there are are 3 or more data frames to merge
\item there are no duplicates in any of the data frames
The function performs a full outer join with base R
merge(df1, df2, by = id_col, all = TRUE)
TO DO: if only two dfs in list, print message or do one merge
}
}
