% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_clean_spread_repeated.R
\name{epi_clean_spread_repeated}
\alias{epi_clean_spread_repeated}
\title{Spread repeated observations}
\usage{
epi_clean_spread_repeated(df = NULL, rep_col = "", id_col_num = 1)
}
\arguments{
\item{df}{A dataframe in long format}

\item{rep_col}{Column index or string to use to base spread on.}

\item{id_col_num}{Column index with ID values. Default is 1.}
}
\value{
a nested list with as many sub-lists as unique values contained in the column passed as rep_col. Column headers are renamed using
\code{\link{epi_clean_add_colname_suffix}} and will contain suffixes taken from
the unique values in rep_col. The ID column passed as id_col_num is included in each sub-list.
}
\description{
Create a single data frame with unique rows (individuals) and repeated observations across columns. A column with the replicate/repeated observation/time-point number for each row must be provided. Visit codes may be zero or nonconsecutive, but they must not be missing and each identifier/visit pair must be unique.
}
\examples{
\dontrun{
n <- 20
df <- data.frame(
  var_id = rep(1:(n / 2), each = 2),
  var_to_rep = rep(c("Pre", "Post"), n / 2),
  x = rnorm(n),
  y = rbinom(n, 1, 0.50),
  z = rpois(n, 2)
)
df
df_spread <- epi_clean_spread_repeated(df, "var_to_rep", 1)
# Returns a nested list:
df_spread
}

}
\seealso{
\code{\link{epi_clean_add_colname_suffix}}, \code{\link{epi_clean_merge_nested_dfs}}, \code{\link{epi_clean_transpose}}
}
\author{
Antonio Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
