% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_clean_transpose.R
\name{epi_clean_transpose}
\alias{epi_clean_transpose}
\title{Transposes a dataframe preserving row and column names}
\usage{
epi_clean_transpose(df = NULL, id_col_num = "")
}
\arguments{
\item{df}{A dataframe object to transpose. Coerced to data.table and returned as data.frame.}

\item{id_col_num}{Index to identify the column with IDs}
}
\value{
A transposed dataframe with the first column containing the transposed values of the column passed as id_col_num.
}
\description{
Transposes a dataframe preserving row and column names. Based on data.table transpose. Assumes there is an id column with unique IDs
}
\examples{
\dontrun{
library(data.table)
# Generate some data:
n <- 20
df <- data.frame(
  var_id = rep(1:(n / 2), each = 2),
  var_to_rep = rep(c("Pre", "Post"), n / 2),
  x = rnorm(n),
  y = rbinom(n, 1, 0.50),
  z = rpois(n, 2)
)
df$id_col <- rownames(df)
df
id_col <- 6
df_t <- epi_clean_transpose(df = df, id_col)
class(df_t)
dim(df)
dim(df_t)
df_t
names(df_t)
}

}
\seealso{
\code{\link{epi_clean_add_colname_suffix}}, \code{\link{epi_clean_spread_repeated}}, \code{\link{epi_clean_merge_nested_dfs}}, \code{\link[data.table]{transpose}}, \code{\link[data.table]{as.data.table}}.
}
\author{
Antonio Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
