% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_clean.R
\name{epi_clean_transpose}
\alias{epi_clean_transpose}
\title{Compare two columns which may have duplicated information
TO DO: test and complete, not working
Test:
Transposes a data frame file and preserves row and column names:
Assumes there is an id column with unique IDs}
\usage{
epi_clean_transpose(df = NULL, id_col_num = "")
}
\description{
Compare two columns which may have duplicated information
TO DO: test and complete, not working
Test:
Transposes a data frame file and preserves row and column names:
Assumes there is an id column with unique IDs
}
