% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_spec.R
\name{epi_eda_spec}
\alias{epi_eda_spec}
\title{Read an EDA specification}
\usage{
epi_eda_spec(path_or_data)
}
\arguments{
\item{path_or_data}{A path to a CSV specification file, or a data frame
containing the specification.}
}
\value{
A validated data frame specification.
}
\description{
Read and validate a specification-first EDA data dictionary.
}
