% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_spec.R
\name{epi_eda_validate_spec}
\alias{epi_eda_validate_spec}
\title{Validate an EDA specification}
\usage{
epi_eda_validate_spec(spec)
}
\arguments{
\item{spec}{A data frame containing an EDA specification.}
}
\value{
The validated specification as a data frame.
}
\description{
Validate the columns and basic values required by a specification-first EDA
data dictionary.
}
