% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_head_and_tail.R
\name{epi_head_and_tail}
\alias{epi_head_and_tail}
\title{Print the first few rows and last few rows of a data frame}
\usage{
epi_head_and_tail(df = NULL, rows = 5, cols = 5, last_cols = FALSE)
}
\arguments{
\item{df}{data.frame as input}

\item{rows}{Number of rows to print, default is 5}

\item{cols}{Number of columns to print, default is 5}

\item{last_cols}{Print the last columns instead of the first few,
default is FALSE}
}
\value{
A data.frame containing the selected head and tail rows and columns.
}
\description{
epi_head_and_tail() prints the first and last few rows of a data.frame
}
\note{
Similar to data.table printing of a data.table object but works for
a tibble or data.frame, or any object that can be coerced to a data.frame
}
\examples{
\dontrun{
n <- 20
df <- data.frame(
  var_id = rep(1:(n / 2), each = 2),
  var_to_rep = rep(c("Pre", "Post"), n / 2),
  x = rnorm(n),
  y = rbinom(n, 1, 0.50),
  z = rpois(n, 2)
)
df
dim(df)
epi_head_and_tail(df)
epi_head_and_tail(df, rows = 2, cols = 2)
epi_head_and_tail(df, rows = 2, cols = 2, last_cols = TRUE)
}

}
\author{
Antonio Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
