% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_list_head.R
\name{epi_list_head}
\alias{epi_list_head}
\title{Print first few rows of each element of a list

epi_list_head() Prettier print of lists}
\usage{
epi_list_head(list = NULL, rows = 3, max = length(list))
}
\arguments{
\item{list}{a list object to print}

\item{rows}{Number of rows to print, default is 3}

\item{max}{Maximum number of elements in list to print, default is the
length of the list passed}
}
\description{
Print first few rows of each element of a list

epi_list_head() Prettier print of lists
}
\examples{

\dontrun{
epi_list_head(as.list(df), 5, 4)
}

}
\author{
Antonio Berlanga-Taylor <\url{https://github.com/AntonioJBT/episcout}>
}
