% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_list_tail.R
\name{epi_list_tail}
\alias{epi_list_tail}
\title{Print the last few rows of each element of a list}
\usage{
epi_list_tail(list = list, rows = 3, max = length(list))
}
\arguments{
\item{list}{a list object to print}

\item{rows}{Number of rows to print, default is 3}

\item{max}{Maximum number of elements in list to print, default is the
length of the list passed}
}
\description{
epi_list_tail() Prettier print of lists for head and tail
}
\examples{

\dontrun{
epi_list_tail(as.list(df), 5, 4)
}

}
\author{
Antonio Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
