% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot_hist.R
\name{epi_plot_hist}
\alias{epi_plot_hist}
\title{Histogram wrapper function using ggplot2}
\usage{
epi_plot_hist(df = NULL, var_x = "", ...)
}
\arguments{
\item{df}{data.frame with x var to plot}

\item{var_x}{Variable to plot on x-axid passed as a string.}

\item{...}{passes arguments to ggplot2::geom_ such as breaks, colour, fill, alpha, etc.}
}
\value{
Prints a ggplot2 histogram
}
\description{
epi_plot_hist()
}
\note{
For other options, save as object and build on the layers.
var_x is passed to ggplot2::aes_string
}
\examples{

\dontrun{



}

}
\seealso{
\code{\link{epi_plot_list}},
\code{\link{epi_plots_to_grid}},
\code{\link{epi_plot_cow_save}}.
}
\author{
Antonio Berlanga-Taylor <\url{https://github.com/AntonioJBT/episcout}>
}
