% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_plot.R
\name{epi_plot_km}
\alias{epi_plot_km}
\title{Plot Kaplan-Meier survival curves}
\usage{
epi_plot_km(
  survfit_obj,
  group_var = NULL,
  facet_var = NULL,
  save_path = NULL,
  ...
)
}
\arguments{
\item{survfit_obj}{A \code{survival::survfit} object.}

\item{group_var}{Optional name of the column to use for colour.
Defaults to the first strata variable if present.}

\item{facet_var}{Optional name of the column to facet by.
Defaults to the second strata variable if present.}

\item{save_path}{Optional file path to save the plot via \code{ggplot2::ggsave}.
If \code{NULL}, the plot is not saved.}

\item{...}{Further arguments passed to \code{ggplot2::ggsave} when saving.}
}
\value{
A ggplot object of the Kaplan-Meier curves.
}
\description{
\code{epi_plot_km()} creates Kaplan-Meier plots using ggplot2 from a
\code{survival::survfit} object.
}
\examples{
\dontrun{
library(survival)
fit <- survfit(Surv(time, status) ~ sex, data = lung)
epi_plot_km(fit, save_path = "km.png")
}

}
