% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot_themes.R
\name{epi_plot_theme_1}
\alias{epi_plot_theme_1}
\title{ggthemes for episcout}
\usage{
epi_plot_theme_1(base_size = 11, base_family = "Times")
}
\value{
None, these functions are called from plotting functions.
}
\description{
Themes for ggplot2 plots for episcout epi_plot_XXX functions.
Most options are hard-coded, if you don't like the themes it is easier to create
one with your preferences or modify directly in a ggplot2 call.
}
\note{
For examples see eg \code{\link{epi_plot_heatmap_triangle}} and others.
If creating your own themes see for instance:
\url{https://cran.r-project.org/web/packages/ggthemes/vignettes/ggthemes.html}{ggthemes},
\url{http://sape.inf.usi.ch/quick-reference/ggplot2/themes}{ggplot2 themes}.
Some colour codes:
\url{http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/}{Colors (ggplot2)},
\url{http://sape.inf.usi.ch/quick-reference/ggplot2/colour}{ggplot2 Quick Reference}
}
\author{
Antonio Berlanga-Taylor <\url{https://github.com/AntonioJBT/episcout}>
}
