% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot_theme_imss.R
\name{epi_plot_theme_imss}
\alias{epi_plot_theme_imss}
\title{IMSS ggplot theme}
\usage{
epi_plot_theme_imss(base_size = 13, base_family = "Times")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme object.
}
\description{
Provides a ggplot2 theme with IMSS corporate colours. Package dependencies are
checked at run time.
}
\examples{
if (requireNamespace('ggplot2', quietly = TRUE) &&
    requireNamespace('ggthemes', quietly = TRUE)) {
  ggplot2::ggplot(mtcars, ggplot2::aes(mpg, wt)) +
    ggplot2::geom_point() +
    epi_plot_theme_imss()
}
}
