% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot_volcano.R
\name{epi_plot_volcano}
\alias{epi_plot_volcano}
\title{Volcano plot}
\usage{
epi_plot_volcano(
  logFC = NULL,
  p_val = NULL,
  main = "",
  pch = 20,
  h_abline = 2,
  v_abline = c(0.8, 1.2),
  ...
)
}
\arguments{
\item{logFC}{Numeric vector of log2 fold changes.}

\item{p_val}{Numeric vector of raw p-values (will be transformed to
\code{-log10(p_val)}).}

\item{main}{Main title passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}

\item{pch}{Plotting character passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}

\item{h_abline}{Horizontal line position passed to \code{\link[graphics:abline]{graphics::abline()}}.}

\item{v_abline}{Vertical line positions passed to \code{\link[graphics:abline]{graphics::abline()}}.}

\item{...}{Additional arguments passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
Invisibly returns \code{NULL} and draws the plot to the active device.
}
\description{
Draw a simple volcano plot given log2 fold changes and raw p-values.
}
\note{
Assumes small effect sizes with vertical cut-offs at 0.8 and 1.2
log2 fold change and a horizontal cut-off corresponding to a p-value of 0.05.
}
\examples{
\dontrun{
log2fc <- rnorm(100)
pvals <- runif(100)
epi_plot_volcano(log2fc, pvals)
}

}
\seealso{
\code{\link[=epi_plot_list]{epi_plot_list()}}, \code{\link[=epi_plots_to_grid]{epi_plots_to_grid()}},
\code{\link[=epi_plot_cow_save]{epi_plot_cow_save()}}, \code{\link[graphics:plot.default]{graphics::plot()}}
}
\author{
Antonio Berlanga-Taylor \url{https://github.com/AntonioJBT/episcout}
}
