% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot_volcano.R
\name{epi_plot_volcano}
\alias{epi_plot_volcano}
\title{Volcano plot}
\usage{
epi_plot_volcano(logFC = NULL, adj.P.Val = NULL, main = "",
  pch = 20, h_abline = 2, v_abline = c(0.8, 1.2), ...)
}
\arguments{
\item{logFC}{limma's estimate of the log2-fold-change.}

\item{adj.P.Val}{adjusted p-value or q-value. This is converted to -log10(adj.P.Val)}

\item{main}{base plot() main parameter. String to use as main title.
Default is an empty string.}

\item{pch}{plotting character, base plot() pch parameter. Default is 20.}

\item{h_abline}{add straight lines, base plot() abline parameter. Default is 2.}

\item{v_abline}{add straight lines, base plot() abline parameter. Default is c(0.8, 1.2).}

\item{...}{Pass any other arguments to base plot()}
}
\description{
epi_plot_volcano()
Designed to take limma's topTable output as input.
}
\note{
Assumes small effect sizes with vertical line cuts at 0.8 and 1.2 logFC.
The horizontal line cuts at a p-value of 0.05.
}
\seealso{
\code{\link{epi_plot_list}},
\code{\link{epi_plots_to_grid}},
\code{\link{epi_plot_cow_save}},
\code{\link[limma]{topTable}},
\code{\link{plot}}.
}
\author{
Antonio Berlanga-Taylor <\url{https://github.com/AntonioJBT/episcout}>
}
