% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_read.R
\name{epi_read}
\alias{epi_read}
\title{Wraps tibble and data.table with several assumptions as defaults}
\usage{
epi_read(input_name = "", na.strings = c(-Inf, "NULL", NULL, ".", " ",
  "NA", "NaN", NA, "<NA>"), header = TRUE, stringsAsFactors = FALSE,
  strip.white = TRUE, ...)
}
\arguments{
\item{input_name}{file to read as a string}

\item{...}{pass any other data.table::fread() parameters

Other parameters as specified by data.table::fread()}
}
\value{
A tibble
}
\description{
epi_read() Read files with a consistent convenience function
}
\examples{
Not run:

\dontrun{super_data <- epi_read('super_data.tsv')}

}
\seealso{
\code{\link[data.table]{fread}}, \code{\link[tibble]{as.tibble}}.
}
\author{
Antonio Berlanga-Taylor <\url{https://github.com/AntonioJBT/episcout}>
}
