% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats_corr_triangle.R
\name{epi_stats_corr_triangle}
\alias{epi_stats_corr_triangle}
\title{Get the lower triangle from a correlation matrix}
\usage{
epi_stats_corr_triangle(cormat = "cormat_all$cormat")
}
\arguments{
\item{cormat}{a matrix object (usually the output of Hmisc::rcorr()
or episcout::epi_stats_corr()). Default 'cormat_all$cormat'}
}
\value{
Returns a list object melted_triangles containing the lower triangle of the
correlation matrix (cormat) with the correlation (cormat_melted_triangle_r) and
p-values (cormat_melted_triangle_pval).
}
\description{
Keep only the lower triangle of the correlation matrix
Useful to create a nicer heatmap. Requires the original, unmelted correlation
matrix with both correlation (r) and p-values (P).
}
\examples{

\dontrun{
df <- data.frame(var_id = rep(1:(n / 2), each = 2),
var_to_rep = rep(c("Pre", "Post"), n / 2),
x = rnorm(n),
y = rbinom(n, 1, 0.50),
z = rpois(n, 2)
)
epi_clean_count_classes(df)
df_corr <- df \%>\%select_if(~ epi_clean_cond_numeric(.))
df_corr <- df_corr[, -1] # exclude var_id
cormat_all <- epi_stats_corr(df_corr, method = 'pearson')
melted_triangles <- epi_stats_corr_triangle(cormat = cormat_all$cormat)
melted_triangles$cormat_melted_triangle_r
melted_triangles$cormat_melted_triangle_pval
class(melted_triangles)
}
}
\seealso{
\code{\link{epi_stats_corr}}, \code{\link{epi_stats_corr_rename}},
\code{\link{epi_plot_heatmap}}, \code{\link{epi_plot_heatmap_triangle}},
\code{\link[Hmisc]{rcorr}}, \code{\link[data.table]{melt}}
}
\author{
Antonio Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
