% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats_dates.R
\name{epi_stats_dates}
\alias{epi_stats_dates}
\title{Calculate Descriptive Date Statistics}
\usage{
epi_stats_dates(date_vector)
}
\arguments{
\item{date_vector}{A vector of dates or datetimes of class \code{Date}, \code{IDate}, \code{POSIXct} or \code{POSIXlt}.}
}
\value{
A \code{data.frame} containing statistics such as N, N Missing, N Unique, Min, 25\%, Median, 75\%, Max, IQR, Most Common, and Range (Days).
}
\description{
Calculates and returns key descriptive statistics for a vector of dates or datetimes, including minimum, maximum, interquartile range (IQR), and quartiles. It is compatible with \code{Date}, \code{IDate}, \code{POSIXct} and \code{POSIXlt} objects.
}
\note{
The default origin for numeric date conversion in R is "1970-01-01".
}
\examples{
sample_dates <- as.Date(c("2020-01-01", "2020-05-15", "2020-12-31", "2021-01-01"))
epi_stats_dates(sample_dates)

}
\seealso{
\code{\link{epi_stats_summary}}, \code{\link{epi_stats_format}}, \code{\link{epi_stats_numeric}}
}
\author{
Antonio J Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
