% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats_dates.R
\name{epi_stats_dates_freq}
\alias{epi_stats_dates_freq}
\title{Date Differences and Monthly Frequencies}
\usage{
epi_stats_dates_freq(date_vector)
}
\arguments{
\item{date_vector}{A vector of dates of class \code{Date} or \code{IDate}.}
}
\value{
A list with two elements:
\describe{
\item{date_differences}{Differences in days between sorted consecutive dates.}
\item{frequencies}{A table of counts by year-month.}
}
}
\description{
Computes differences between consecutive dates and frequency of
observations by year-month for a vector of dates.
}
\examples{
sample_dates <- as.Date(c("2020-01-01", "2020-01-03", "2020-02-01"))
epi_stats_dates_freq(sample_dates)

}
