% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats_dates.R
\name{epi_stats_dates_multi}
\alias{epi_stats_dates_multi}
\title{Summarize multiple date columns (Wide Format)}
\usage{
epi_stats_dates_multi(df)
}
\arguments{
\item{df}{A dataframe containing multiple date columns}
}
\value{
A wide-format tibble summarizing date statistics
}
\description{
Summarize multiple date columns (Wide Format)
}
