% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats_sum_vars.R
\name{epi_stats_factors}
\alias{epi_stats_factors}
\title{Summarize Factor Variables}
\usage{
epi_stats_factors(df)
}
\arguments{
\item{df}{A dataframe containing factor variables}
}
\value{
A tibble summarizing factor variables in wide format
}
\description{
Summarize Factor Variables
}
