% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats_format.R
\name{epi_stats_format}
\alias{epi_stats_format}
\title{Format a dataframe with numerical or integer columns}
\usage{
epi_stats_format(df = NULL, skip = NULL, digits = 2, ...)
}
\arguments{
\item{df}{Data.frame with summary to clean up.}

\item{skip}{Columns to skip, pass as a string. Default is NULL.}

\item{digits}{Number of digits to print. Default is 2.}

\item{...}{Any other parameter that can be passed to round().}
}
\value{
A data.frame with formatted and rounded values.
}
\description{
epi_stats_format() formats columns so that digits appear
even if they are x.00. Useful for saving a table with descriptive statistics.
A data frame with an id column is expected. Check if column
is numeric or integer, other types are not formatted. Pass a vector
of column number to skip if needed. Values are rounded to the option
passed to 'digits' digits = 2 by default
}
\seealso{
\code{\link{epi_stats_summary}},
\code{\link{epi_stats_tidy}},
\code{\link{epi_clean_cond_numeric}},
\code{\link[base]{format}},
\code{\link[base]{round}}.
}
\author{
Antonio J Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
