% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats_numeric.R
\name{epi_stats_numeric}
\alias{epi_stats_numeric}
\title{Get summary descriptive statistics for numeric columns.}
\usage{
epi_stats_numeric(num_vec = NULL, na.rm = TRUE, coef = 1.5, ...)
}
\arguments{
\item{num_vec}{Numeric vector to test.}

\item{na.rm}{Remove NA values, default is TRUE}

\item{coef}{Coefficient for outlier detection, default is 1.5}

\item{...}{is passed to skewness() and kurtosis()}
}
\value{
Data.frame with results from tests performed
}
\description{
epi_stats_numeric() generates several descriptive summary statistics for numeric
variables.
}
\note{
Normality is tested with Shapiro-Wilk (small values indicate
non-normality). Testing normality can be contentious, likely uninformative
and with Shapiro-Wilk can only be done for sample size between 3-5000.
The package e1071 is used for skewness and kurtosis.
For skewness: negative/longer left tail, positive/longer right tail,
values above 1 usually means non-normality. For kurtosis consider lower values,
broader shape and longer tails (platy ~<3), normal (meso ~3)
and slender/no tails (lepto ~<3). Outliers are detected with the Tukey
method (above and below 1.5 * IQR). na.rm is TRUE by default for all tests.
}
\seealso{
\code{\link{epi_stats_summary}},
\code{\link{epi_stats_format}},
\code{\link{epi_stats_tidy}},
\code{\link[e1071]{skewness}},
\code{\link[e1071]{kurtosis}}.
}
\author{
Antonio J Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
