% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats_prop_outcome.R
\name{epi_stats_prop_outcome}
\alias{epi_stats_prop_outcome}
\title{Calculate the proportion of individuals who suffered an event
at a given point in time}
\usage{
epi_stats_prop_outcome(
  df,
  outcome_var_window,
  pop_at_risk_var,
  analysis_window,
  round_dig = 4
)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{outcome_var_window}{A string representing the column name of the outcome variable at the analysis window. Expects an integer where 1 = event, 0 = no event.}

\item{pop_at_risk_var}{A string representing the column name of the population at risk variable. This is matched against the analysis window to obtain a subset.}

\item{analysis_window}{A value from the variable passed as \code{pop_at_risk_var} representing the analysis window.}

\item{round_dig}{An integer representing the number of decimal places to round the result to. Default is 4.}
}
\value{
The proportion of individuals who suffered the event at the given point, printed and returned as a numeric value.
}
\description{
It can be used to estimate the case fatality rate.
}
\examples{
\dontrun{
df <- data.frame(
  d_T0_outcome = c(1, 0, 1, 0, 1),
  d_time_cuts_prev = c('T0', 'T1', 'T0', 'T1', 'T0')
)
outcome_var_window <- 'd_T0_outcome'
pop_at_risk_var <- 'd_time_cuts_prev'
analysis_window <- 'T0'
epi_stats_prop_outcome(df, outcome_var_window, pop_at_risk_var, analysis_window)
}
}
