% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats_summary.R
\name{epi_stats_summary}
\alias{epi_stats_summary}
\title{Get summary statistics from a data frame with multiple columns}
\usage{
epi_stats_summary(
  df = NULL,
  codes = NULL,
  class_type = "chr_fct",
  action = "exclude"
)
}
\arguments{
\item{df}{Data frame}

\item{codes}{Specify codes to summarise or exclude as string. Default is NULL.}

\item{class_type}{Class of variables to summarise, 'chr_fct' or 'int_num'.
Default is character and factor.}

\item{action}{Values to summarise, 'codes_only' or 'exclude'. Default is 'exclude'.}
}
\value{
A data.frame as tibble with summaries.
}
\description{
epi_stats_summary() provides summary descriptive statistics
for columns belonging to either character and factor (class_type = 'chr_fct')
or integer and numeric (class_type = 'int_num') while discarding values
provided (codes). This is useful if data frame has contingency codes.
Columns are ordered according to order in contingency codes option.
Rows are then ordered in decreasing order according to column provided.
}
\note{
Desgined with data frames that require pre-processing and likely
have contingency and database codes. Action 'exclude' excludes the string values
provided from the summary. Useful to quickly assess what a data.frame contains,
types of values in each column and summary statistics if excluding codes.
}
\seealso{
\code{\link{epi_stats_numeric}},
\code{\link{epi_stats_format}},
\code{\link{epi_stats_tidy}},
\code{\link{epi_clean_cond_chr_fct}},
\code{\link{epi_clean_cond_numeric}}.
}
\author{
Antonio J Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
