% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats_tidy.R
\name{epi_stats_tidy}
\alias{epi_stats_tidy}
\title{Tidy up a data.frame with summary values}
\usage{
epi_stats_tidy(
  sum_df = NULL,
  order_by = "percent",
  perc_n = NULL,
  digits = 2,
  decreasing = TRUE
)
}
\arguments{
\item{sum_df}{Data.frame with summary to clean up.}

\item{order_by}{Column to order results by. Default is 'percent'.}

\item{perc_n}{Number of rows from original dataframe to calculate percentage. Must
be passed manually.}

\item{digits}{= 2,}

\item{decreasing}{Sort values by decreasing order. Default is TRUE.}
}
\value{
Returns a dataframe as a tibble with values ordered and spread.
Adds row sums and percentage.
}
\description{
epi_stats_tidy() cleans up the output from epi_stats_summary() and
epi_stats_numeric(). Values are rounded to digits (default is 2).
format(x, nsmall = digits) is used to ensure xx.00 are printed.
Ordering uses as.numeric(as.character(x)) as 'percent' or other numeric
column is assumed to be the preferred option. 'decreasing' is passed to order.
}
\note{
Note that format() will likely change the class type.
}
\seealso{
\code{\link{epi_stats_summary}},
\code{\link{epi_stats_format}},
\code{\link{epi_stats_numeric}}.
}
\author{
Antonio J Berlanga-Taylor <\url{https://github.com/AntonioJBT/episcout}>
}
