% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_utils_log.R
\name{epi_utils_log}
\alias{epi_utils_log}
\title{Save a log of an R session}
\usage{
epi_utils_log(output_prefix = NULL)
}
\arguments{
\item{output_prefix}{File name as string, '_log.txt' is appended
to a log file. Default is e.g.
'session_date_log.txt' eg 'session_2019-03-01_log.txt'}
}
\value{
Saves a log file to disk
}
\description{
epi_utils_log() is a convenience function to log details of an R
session.
}
\examples{

\dontrun{

x <- stats::runif(20)
y <- list(a = 1, b = TRUE, c = "oops")
# See objects to save:
ls()
objects_to_save <- c('x', 'y')
epi_utils_session(output_prefix = 'xy',
                  objects_to_save = objects_to_save
                 )
epi_utils_log('xy')
}

}
\seealso{
\code{\link[utils]{sessionInfo}},
\code{\link[episcout]{epi_utils_session}}.
}
\author{
Antonio J Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
