% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_utils_session.R
\name{epi_utils_session}
\alias{epi_utils_session}
\title{Save an R session}
\usage{
epi_utils_session(output_prefix = NULL, objects_to_save = NULL, ...)
}
\arguments{
\item{output_prefix}{File name as string, '.RData' is appended
to saved objects. Default is e.g.
'session_date.RData' eg 'session_2019-03-01.RData'}

\item{objects_to_save}{String with list of objects to save. Assumes you only
want a subset from the R session. Use saveRDS() for one object or save.image()
for the whole session instead.}

\item{...}{pass any other parameters from save() such as compress = 'gzip'}
}
\value{
Saves an external representation of the R objects to file
}
\description{
epi_utils_session() is a convenience function to save an R
session.
}
\examples{

\dontrun{

x <- stats::runif(20)
y <- list(a = 1, b = TRUE, c = "oops")
# See objects to save:
ls()
objects_to_save <- c('x', 'y')
epi_utils_session(output_prefix = 'xy',
                  objects_to_save = objects_to_save
                 )
dir()
rm(list = objects_to_save)
ls()
load('xy.RData')
ls()
}

}
\seealso{
\code{\link[base]{save.image}},
\code{\link[base]{saveRDS}},
\code{\link[base]{load}},
\code{\link[base]{readRDS}},
\code{\link[episcout]{epi_utils_log}}.
}
\author{
Antonio J Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
