% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_write.R
\name{epi_write}
\alias{epi_write}
\title{Write files with a consistent convenience function}
\usage{
epi_write(
  file_object = NULL,
  file_name = "my_file.tsv",
  row.names = FALSE,
  quote = FALSE,
  sep = "\\t",
  na = "NA",
  col.names = TRUE,
  ...
)
}
\arguments{
\item{file_object}{Object to write}

\item{file_name}{file name as a string, default is 'my_file.tsv'}

\item{row.names}{as for data.table::fwrite(), default is also FALSE}

\item{quote}{as for data.table::fwrite(), default is FALSE}

\item{sep}{as for data.table::fwrite(), default is tab separated}

\item{na}{as for data.table::fwrite(), default is 'NA'}

\item{col.names}{as for data.table::fwrite(), default is also TRUE}

\item{...}{pass any other data.table::fwrite() parameters}
}
\value{
None
}
\description{
epi_write() writes files with a consistent convenience function.
Wraps data.table's fwrite with several assumptions as defaults.
}
\note{
Other parameters as specified by data.table::fwrite()
You are probably better off using the standard write.csv(),
data.table::fwrite() directly or other functions unless you are
writing out several similarly constructed objects.
Files are written with data.table. Column names are written,
row names are not, quotes are not used. Columns are tab separated.
NAs are written as 'NA'. Assumes the first row in the object is a header.
}
\examples{

\dontrun{
epi_write(some_dataframe, 'some_dataframe.tsv')
}

}
\seealso{
\code{\link[data.table]{fwrite}}
}
\author{
Antonio Berlanga-Taylor \if{html}{\out{<\url{https://github.com/AntonioJBT/episcout}>}}
}
