% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_write_df.R
\name{epi_write_df}
\alias{epi_write_df}
\title{Save a Data Frame to a File with Custom Formatting}
\usage{
epi_write_df(df, results_subdir, file_n, suffix)
}
\arguments{
\item{df}{A data frame to be written to the file.}

\item{results_subdir}{A character string specifying the directory where the file will be saved.}

\item{file_n}{A character string specifying the base name of the file (without the suffix).}

\item{suffix}{A character string specifying the file extension (e.g., "txt", "csv").}
}
\value{
The full file path of the saved file.
Prints a message indicating the file's location.
}
\description{
This function constructs a file path using a specified subdirectory, file name, and suffix, then writes a data frame to that file. It is designed for flexibility in saving results with standardized naming conventions.
}
\examples{
\dontrun{
# Example usage
results_subdir <- "output"
file_n <- "desc_dates"
suffix <- "txt"
epi_write_df(sum_dates_df, results_subdir, file_n, suffix)
}

}
