% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episcout-package.R
\docType{package}
\name{episcout-package}
\alias{episcout}
\alias{episcout-package}
\title{episcout: Quickly Clean, Explore and Visualise Large Epidemiological Datasets}
\description{
Facilitates cleaning, exploring and visualising large-ish epidemiological datasets (hundreds of thousands to millions of observations with tens to hundreds of variables). Provides helper functions for preprocessing and descriptive work, plus specification-first exploratory data analysis workflows for epidemiological datasets. Assumes you're happy with tidyverse and the basics of data.table.
}
\author{
\strong{Maintainer}: Antonio Berlanga-Taylor \email{antoniojberlanga@gmail.com}

Authors:
\itemize{
  \item Antonio Berlanga-Taylor \email{antoniojberlanga@gmail.com}
}

}
\keyword{internal}
