% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episcout-package.R
\docType{package}
\name{episcout-package}
\alias{episcout}
\alias{episcout-package}
\title{episcout: episcout}
\description{
Facilitates cleaning, exploring and visualising large-ish datasets (hundreds of thousands to millions of observations with tens to hundreds of variables). These are mostly wrapper and convenience functions to pre-process (wrangle, explore, clean, etc.) data-sets. Assumes you're happy with tidyverse and the basics of data.table.
}
\author{
\strong{Maintainer}: Antonio Berlanga-Taylor \email{a.berlanga@imperial.ac.uk}

}
\keyword{internal}
