% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot_colour_palettes.R
\docType{data}
\name{palette_IMSS_accessible}
\alias{palette_IMSS_accessible}
\title{Colour-blind–friendly IMSS palette}
\format{
An object of class \code{character} of length 10.
}
\usage{
palette_IMSS_accessible
}
\value{
A character vector of hex colour codes.
}
\description{
A ten‐hex vector that’s easier to distinguish for viewers
with colour‐vision deficiencies.
}
\keyword{datasets}
